Move NPC autocal code out to a separate class.
